refactor(components): change components directory structure
