+++ web/src/components/pages/ProfilePage.tsx
@@ -1,11 +1,11 @@
 import { useDispatch, useSelector } from "react-redux";
-import { AppDispatch, RootState } from "../../store";
-import { logoutUser, deleteUser } from "../reducers/userReducer";
+import { AppDispatch, RootState } from "../../../store";
+import { logoutUser, deleteUser } from "../../reducers/userReducer";
 import { useNavigate, Link } from "react-router-dom";
-import { UserType } from "../types";
-import TutorProfile from "./TutorProfile";
-import StudentForm from "./StudentForm";
-import StudentProfile from "./StudentProfile";
+import { UserType } from "../../types";
+import TutorProfile from "../TutorProfile";
+import StudentForm from "../StudentForm";
+import StudentProfile from "../StudentProfile";
 
 const ProfilePage = () => {
     const user = useSelector((state: RootState) => state.user);
